requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.10.0
async-timeout==5.0.1
billiard==4.2.3
brotli==1.2.0
cachetools==6.2.2
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2025.6
cssselect2==0.8.0
django==4.2.25
django-crispy-forms==2.4
django-simple-history==3.10.1
fonttools==4.60.1
gevent==21.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==1.1.3.post0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.11
kombu==5.5.4
numpy==2.0.2
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==11.3.0
prompt-toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.26.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydyf==0.11.0
pyparsing==3.2.5
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
redis==7.0.1
requests==2.32.5
rsa==4.9.1
scipy==1.13.1
six==1.17.0
sqlparse==0.5.3
statsmodels==0.14.5
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
zope-event==6.0
zope-interface==8.0.1
zopfli==0.2.3.post1
